turning on testing mode and starting refactor phase 1
