+++ .github/actions/analyze-workflow-data/analyze-workflow-data.js
@@ -11,1237 +11,67 @@ const github = require('@actions/github'); // GitHub API client
 const fs = require('fs'); // File system operations
 const path = require('path'); // File path utilities
 
-// Constants
-const DEFAULT_LOOKBACK_DAYS = 15;
-const SHA_SHORT_LENGTH = 7;
-const SUCCESS_RATE_DECIMAL_PLACES = 2;
-const SUCCESS_EMOJI = '✅';
-const FAILURE_EMOJI = '❌';
-const EMPTY_VALUE = '—';
-// Default owner override when a test name cannot be extracted
-const DEFAULT_INFRA_OWNER = { id: 'S0985AN7TC5', name: 'metal infra team' };
+// Import modules
+const dataLoading = require('./lib/data-loading');
+const errorProcessing = require('./lib/error-processing');
+const analysis = require('./lib/analysis');
+const reporting = require('./lib/reporting');
+
+// Import constants used in this file
+const {
+  DEFAULT_LOOKBACK_DAYS,
+  SHA_SHORT_LENGTH,
+  EMPTY_VALUE,
+  DEFAULT_INFRA_OWNER,
+} = dataLoading;
+// Import functions from modules
+const {
+  loadAnnotationsIndexFromFile,
+  getAnnotationsDirForRunId,
+  loadLogsIndexFromFile,
+  getGtestLogsDirForRunId,
+  getOtherLogsDirForRunId,
+  loadLastSuccessTimestamps,
+  getTimeSinceLastSuccess,
+  getCommitDescription,
+  loadCommitsIndex,
+  listCommitsBetweenOffline,
+  setAnnotationsIndexMap,
+  setGtestLogsIndexMap,
+  setOtherLogsIndexMap,
+  setLastSuccessTimestamps,
+  setCommitsIndex,
+} = dataLoading;
+
+const {
+  loadOwnersMapping,
+  normalizeOwners,
+  extractSignificantTokens,
+  getJobNameComponentTail,
+  findOwnerForLabel,
+  inferJobAndTestFromSnippet,
+  resolveOwnersForSnippet,
+  fetchErrorSnippetsForRun,
+  renderErrorsTable,
+} = errorProcessing;
+
+const {
+  getWorkflowStats,
+  findFirstFailInWindow,
+  renderCommitsTable,
+  fetchPRInfo,
+  fetchCommitAuthor,
+} = analysis;
+
+const {
+  getWorkflowLink,
+  findGoodBadCommits,
+  getLastRunInfo,
+  generateSummaryBox,
+  buildReport,
+  filterRunsByDate,
+} = reporting;
 
-// Optional annotations index mapping (runId -> directory)
-// This is populated when action inputs provide their paths.
-let __annotationsIndexMap = undefined;
-
-// Optional gtest logs index mapping (runId -> directory)
-let __gtestLogsIndexMap = undefined;
-
-// Optional other logs index mapping (runId -> directory)
-let __otherLogsIndexMap = undefined;
-
-// Optional last success timestamps mapping (workflow name -> timestamp info)
-let __lastSuccessTimestamps = undefined;
-
-
-
-/**
- * Resolve the extracted logs directory for a given run ID using the loaded index.
- */
-function loadAnnotationsIndexFromFile(filePath) {
-  try {
-    if (!filePath || !fs.existsSync(filePath)) return undefined;
-    const raw = fs.readFileSync(filePath, 'utf8');
-    const json = JSON.parse(raw);
-    const map = new Map();
-    if (json && typeof json === 'object' && !Array.isArray(json)) {
-      for (const [k, v] of Object.entries(json)) {
-        if (typeof v === 'string' && v) map.set(String(k), v);
-      }
-      return map;
-    }
-  } catch (_) { /* ignore */ }
-  return undefined;
-}
-
-function getAnnotationsDirForRunId(runId) {
-  try {
-    if (!__annotationsIndexMap) return undefined;
-    const key = String(runId);
-    return __annotationsIndexMap.get(key);
-  } catch (_) {
-    return undefined;
-  }
-}
-
-function loadLogsIndexFromFile(filePath) {
-  try {
-    if (!filePath || !fs.existsSync(filePath)) return undefined;
-    const raw = fs.readFileSync(filePath, 'utf8');
-    const json = JSON.parse(raw);
-    const map = new Map();
-    if (json && typeof json === 'object' && !Array.isArray(json)) {
-      for (const [k, v] of Object.entries(json)) {
-        if (typeof v === 'string' && v) map.set(String(k), v);
-      }
-      return map;
-    }
-  } catch (_) { /* ignore */ }
-  return undefined;
-}
-
-function getGtestLogsDirForRunId(runId) {
-  try {
-    if (!__gtestLogsIndexMap) return undefined;
-    const key = String(runId);
-    return __gtestLogsIndexMap.get(key);
-  } catch (_) {
-    return undefined;
-  }
-}
-
-function getOtherLogsDirForRunId(runId) {
-  try {
-    if (!__otherLogsIndexMap) return undefined;
-    const key = String(runId);
-    return __otherLogsIndexMap.get(key);
-  } catch (_) {
-    return undefined;
-  }
-}
-
-function loadLastSuccessTimestamps(filePath) {
-  try {
-    if (!filePath || !fs.existsSync(filePath)) return undefined;
-    const raw = fs.readFileSync(filePath, 'utf8');
-    const json = JSON.parse(raw);
-    const map = new Map();
-    if (json && typeof json === 'object' && !Array.isArray(json)) {
-      for (const [k, v] of Object.entries(json)) {
-        if (v && typeof v === 'object') map.set(String(k), v);
-      }
-      return map;
-    }
-  } catch (_) { /* ignore */ }
-  return undefined;
-}
-
-function getTimeSinceLastSuccess(workflowName) {
-  try {
-    if (!__lastSuccessTimestamps) return EMPTY_VALUE;
-    const info = __lastSuccessTimestamps.get(workflowName);
-    if (!info) return EMPTY_VALUE;
-
-    if (info.never_succeeded) {
-      return 'Never';
-    }
-
-    if (!info.timestamp) return EMPTY_VALUE;
-
-    const lastSuccessDate = new Date(info.timestamp);
-    const now = new Date();
-    const daysSince = Math.floor((now - lastSuccessDate) / (1000 * 60 * 60 * 24));
-
-    if (daysSince === 0) return 'Today';
-    if (daysSince === 1) return '1 day ago';
-    return `${daysSince} days ago`;
-  } catch (_) {
-    return EMPTY_VALUE;
-  }
-}
-
-// CODE FOR OWNERSHIP MAPPING
-
-// Owners mapping cache
-let __ownersMapping = undefined; // Stores the owners data
-function loadOwnersMapping() {
-  if (__ownersMapping !== undefined) return __ownersMapping; // If the owners mapping is already loaded, return it
-  try {
-    const ownersPath = path.join(__dirname, 'owners.json'); // Path to the owners.json file
-    if (fs.existsSync(ownersPath)) {
-      const raw = fs.readFileSync(ownersPath, 'utf8'); // Read the owners.json file in some raw format
-      __ownersMapping = JSON.parse(raw); // Parse the raw data into a JSON object
-    } else {
-      __ownersMapping = null; // If the owners.json file does not exist, set the owners mapping to null
-    }
-  } catch (_) {
-    __ownersMapping = null; // If there is an error, set the owners mapping to null
-  }
-  return __ownersMapping;
-}
-
-// Normalizes the owner information to a consistent format for easier processing later
-// Example: ["charlie", { id: "david", name: "David" }] → [{ id: "charlie" }, { id: "david", name: "David" }]
-function normalizeOwners(value) {
-  if (!value) return undefined;
-  // Single string -> id-only
-  if (typeof value === 'string') return [{ id: value }];
-  // Object with id/name
-  if (typeof value === 'object' && !Array.isArray(value)) {
-    const id = value.id || undefined;
-    const name = value.name || undefined;
-    if (!id && !name) return undefined;
-    return [{ id, name }];
-  }
-  // Array -> map each entry
-  if (Array.isArray(value)) {
-    const arr = [];
-    for (const entry of value) {
-      if (typeof entry === 'string') arr.push({ id: entry });
-      else if (entry && typeof entry === 'object') arr.push({ id: entry.id, name: entry.name });
-    }
-    return arr.length ? arr : undefined;
-  }
-  return undefined;
-}
-
-function extractSignificantTokens(str) {
-  if (typeof str !== 'string') return [];
-  return str
-    .split(/\s+[^A-Za-z0-9\s]+\s+/) // Splits the string where there's whitespace plus punctuation plus whitespace
-                                    // For example: "test::function_name" → ["test", "function_name"]
-
-    .flatMap(segment => segment.split(/[\r\n]+/)) // Splits the string where there's a newline
-                                                // For example: "test\nfunction_name" → ["test", "function_name"]
-
-    .map(segment => segment.trim()) // Trims the string
-                                    // For example: "   test" → "test"
-
-    .filter(Boolean) // Removes empty values
-
-    .map(segment => segment.replace(/^[^A-Za-z0-9_\-\s]+|[^A-Za-z0-9_\-\s]+$/g, '').trim()) // Removes non-alphanumeric characters from the beginning and end of each segment
-                                                                                            // Keeps letters, numbers, underscores, hyphens, and spaces
-
-    .filter(segment => segment.replace(/\s+/g, '').length > 1); // Removes segments that are 1 character or less without spaces
-}
-
-function getJobNameComponentTail(component) {
-  const tokens = extractSignificantTokens(component);
-  return tokens.length ? tokens[tokens.length - 1] : undefined; // Returns the last token in the array
-}
-
-function findOwnerForLabel(label) {
-  try {
-    const mapping = loadOwnersMapping(); // Load mapping
-    if (!mapping) return undefined; // If the mapping is not loaded, return undefined
-    const lbl = typeof label === 'string' ? label : ''; // If the label is not a string, return an empty string
-    const labelTokens = extractSignificantTokens(lbl); // Extract significant tokens from the label
-
-    if (Array.isArray(mapping.contains)) { // If the mapping's contains parameter is an array
-      for (const entry of mapping.contains) {
-        // Backward-compat: accept both "job-name-component" and legacy "needle" keys
-        const component = (entry && typeof entry["job-name-component"] === 'string' && entry["job-name-component"]) || (entry && typeof entry.needle === 'string' && entry.needle) || undefined;
-        if (!component) continue; // skip if missing
-        if (lbl.includes(component)) { // direct substring match
-          return normalizeOwners(entry.owner);
-        }
-        // Fuzzy match: try last token from the component
-        const tail = getJobNameComponentTail(component);
-        if (tail && labelTokens.includes(tail)) {
-          return normalizeOwners(entry.owner);
-        }
-        // Additional heuristic: if label tokens end with the last two tokens of the component
-        const componentTokens = extractSignificantTokens(component);
-        if (componentTokens.length >= 2 && labelTokens.length >= 2) {
-          const componentTailPair = componentTokens.slice(-2).join(' ');
-          const labelTailPair = labelTokens.slice(-2).join(' ');
-          if (componentTailPair === labelTailPair) {
-            return normalizeOwners(entry.owner);
-          }
-        }
-      }
-    }
-  } catch (_) {
-    // ignore
-  }
-  return undefined;
-}
-
-// END OF CODE FOR OWNERSHIP MAPPING
-
-// START OF CODE FOR ERROR HANDLING
-
-// Helper: infer job/test from a snippet's existing fields, label, and message
-function inferJobAndTestFromSnippet(snippet) {
-  try {
-    const rawLabel = (snippet && snippet.label) ? String(snippet.label) : '';
-    const rawSnippet = (snippet && snippet.snippet) ? String(snippet.snippet) : '';
-    let jobName = (snippet && typeof snippet.job === 'string') ? snippet.job : '';
-    let testName = (snippet && typeof snippet.test === 'string') ? snippet.test : '';
-
-    // 1) Parse from label when needed
-    if (!jobName || !testName) {
-      if (rawLabel && rawLabel.includes(':')) {
-        const idx = rawLabel.indexOf(':');
-        const left = rawLabel.slice(0, idx).trim();
-        const right = rawLabel.slice(idx + 1).trim();
-        if (!jobName) jobName = left;
-        if (!testName && right) testName = right.replace(/\s*\[[^\]]+\]\s*$/, '').trim();
-      }
-    }
-
-    // 2) Fallback heuristics based on snippet body
-    if (!testName) {
-      const trimmed = rawSnippet.replace(/^\s+/, '');
-      const generic = /(lost\s+connection|timeout|timed\s*out|connection\s+reset|network\s+is\s+unreachable|no\s+space\s+left|killed\s+by|out\s+of\s+memory)/i;
-      const ib = trimmed.indexOf('[');
-      const ispace = trimmed.indexOf(' ');
-      const endWord = (ispace === -1 && ib === -1) ? trimmed.length : Math.min(...[ispace, ib].filter(v => v !== -1));
-      const firstToken = trimmed.slice(0, Math.max(0, endWord)).trim();
-      if (generic.test(trimmed)) testName = 'NA';
-      else if (firstToken && /test/i.test(firstToken)) testName = firstToken;
-      else if (ib !== -1) testName = trimmed.slice(0, ib).trim() || 'NA';
-      else testName = 'NA';
-    }
-
-    // 3) Final fallback for job from label
-    if (!jobName) jobName = rawLabel || '';
-    return { job: jobName.replace(/\s*\[[^\]]+\]\s*$/, ''), test: testName };
-  } catch (_) {
-    return { job: (snippet && snippet.job) || '', test: (snippet && snippet.test) || 'NA' };
-  }
-}
-
-// Helper: resolve owners for a snippet after test has been inferred
-function resolveOwnersForSnippet(snippet, workflowName) {
-  try {
-    // Do not set owners here if already present; we recompute from mapping
-    const label = (snippet && snippet.label) ? String(snippet.label) : '';
-    const stripBracketSuffix = (s) => (typeof s === 'string' ? s.replace(/\s*\[[^\]]+\]\s*$/, '').trim() : s);
-    const cleaned = stripBracketSuffix(label || '');
-    let ownersArr = findOwnerForLabel(cleaned) || findOwnerForLabel(label || '') || [];
-    // Normalize and dedupe
-    const normalized = [];
-    const seenOwners = new Set();
-    for (const o of (ownersArr || [])) {
-      if (!o) continue;
-      const id = o.id || undefined;
-      const name = o.name || undefined;
-      const k = `${id || ''}|${name || ''}`;
-      if (seenOwners.has(k)) continue;
-      seenOwners.add(k);
-      normalized.push({ id, name });
-    }
-    // Apply infra policy: if no owners or test missing/NA, infra is sole owner
-    const testName = (snippet && snippet.test) ? snippet.test : 'NA';
-    const hadOriginalOwners = normalized.length > 0;
-    if (!testName || testName === 'NA') {
-      // Missing/NA test: infra is the only owner; preserve original owners for report context
-      snippet.owner = [DEFAULT_INFRA_OWNER];
-      snippet.owner_source = hadOriginalOwners ? 'infra_due_to_missing_test' : 'infra_due_to_missing_test_no_original';
-      if (hadOriginalOwners) snippet.original_owners = normalized;
-    } else if (!hadOriginalOwners) {
-      // No mapping owners found: default to infra only
-      snippet.owner = [DEFAULT_INFRA_OWNER];
-      snippet.owner_source = 'infra_due_to_no_owner';
-    } else {
-      // Normal case: keep resolved owners
-      snippet.owner = normalized;
-      snippet.owner_source = 'resolved_mapping';
-    }
-  } catch (_) {
-    snippet.owner = [DEFAULT_INFRA_OWNER];
-    snippet.owner_source = 'infra_due_to_error';
-  }
-}
-
-function renderErrorsTable(errorSnippets) {
-  if (!Array.isArray(errorSnippets) || errorSnippets.length === 0) {
-    return '<p><em>No error info found</em></p>';
-  }
-  let __maxJobLen = 0;
-  const rows = errorSnippets.map(obj => {
-    const rawLabel = (obj && obj.label) ? String(obj.label) : '';
-    const rawSnippet = (obj && obj.snippet) ? String(obj.snippet) : '';
-    // Rendering-only: rely on precomputed job/test/owner; do minimal fallbacks for display
-    let jobName = (obj && typeof obj.job === 'string') ? obj.job : '';
-    let testName = (obj && typeof obj.test === 'string') ? obj.test : '';
-    if (!jobName) jobName = rawLabel || '';
-    if (!testName) testName = 'NA';
-    const errorForDisplay = rawSnippet;
-
-    // Aesthetics: drop trailing bracketed status like [failure] or [error]
-    jobName = jobName.replace(/\s*\[[^\]]+\]\s*$/, '');
-    // Prevent excessive wrapping: use non-breaking hyphens and keep slash groups together
-    const jobDisplay = String(jobName)
-      .replace(/-/g, '\u2011')
-      .replace(/\s\/\s/g, '\u00A0/\u00A0');
-    // HTML escape the content
-    const escapeHtml = (str) => String(str).replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;').replace(/'/g, '&#039;');
-    const jobEsc = escapeHtml(jobDisplay).replace(/\r?\n/g, ' ⇥ ');
-
-
-    // THE FOLLOWING SECTION IS PURELY FOR DISPLAY AESTHETICS. NO IMPORTANT LOGIC IS HERE
-    // ========================================================
-    // Build a forced two-line HTML for the job label by inserting a <br/> near the midpoint.
-    // Preference: split at '/' boundary; fallback to nearest space; otherwise hard split at midpoint.
-    let jobHtml = jobEsc;
-    try {
-      const plain = String(jobDisplay).replace(/\r?\n/g, ' ').trim();
-      const mid = Math.floor(plain.length / 2);
-      let breakIdx = -1;
-      // Prefer splitting at slash
-      const slashIdx = plain.indexOf('/');
-      if (slashIdx !== -1) {
-        breakIdx = slashIdx + 1; // keep slash at end of first line
-      } else {
-        // Look for a breakable space near the midpoint (normal or NBSP)
-        const window = 25;
-        const right = plain.slice(mid, Math.min(plain.length, mid + window)).search(/[\u00A0\s]/);
-        const left = plain.slice(Math.max(0, mid - window), mid).lastIndexOf(' ');
-        const leftNbsp = plain.slice(Math.max(0, mid - window), mid).lastIndexOf('\u00A0');
-        const leftBest = Math.max(left, leftNbsp);
-        if (right !== -1) breakIdx = mid + right + 1;
-        else if (leftBest !== -1) breakIdx = Math.max(1, Math.max(0, mid - window) + leftBest + 1);
-      }
-      if (breakIdx <= 0 || breakIdx >= plain.length) breakIdx = Math.max(1, mid);
-      const part1 = plain.slice(0, breakIdx).trimEnd();
-      const part2 = plain.slice(breakIdx).trimStart();
-      jobHtml = `${escapeHtml(part1).replace(/\s/g, '\u00A0')}<br/>${escapeHtml(part2).replace(/\s/g, '\u00A0')}`;
-    } catch (_) { /* keep fallback jobEsc */ }
-    // Track longest job (by characters shown) to size the column later (2-line target ≈ len/2ch)
-    try { __maxJobLen = Math.max(__maxJobLen, jobDisplay.length); } catch (_) { /* ignore */ }
-    const testEsc = escapeHtml(testName).replace(/\r?\n/g, ' ⇥ ');
-    const snippetOneLine = escapeHtml(errorForDisplay || '').replace(/\r?\n/g, ' ⇥ ');
-    let ownerDisplay = 'no owner found';
-    if (obj && Array.isArray(obj.owner) && obj.owner.length) {
-      const names = obj.owner.map(o => (o && (o.name || o.id)) || '').filter(Boolean);
-      if (names.length) ownerDisplay = names.join(', ');
-      // If owner is infra-only due to missing test, surface original pipeline owner for human context
-      if (obj.owner_source && String(obj.owner_source).startsWith('infra_due_to_missing_test') && Array.isArray(obj.original_owners) && obj.original_owners.length) {
-        const origNames = obj.original_owners.map(o => (o && (o.name || o.id)) || '').filter(Boolean);
-        if (origNames.length) ownerDisplay = `${DEFAULT_INFRA_OWNER.name} (pipeline owner: ${origNames.join(', ')})`;
-      }
-    }
-    const ownerEsc = escapeHtml(ownerDisplay);
-    // Force exactly two lines: compute a break point (prefer comma, else space near middle), NBSP around words
-    let ownerHtml = ownerEsc;
-    try {
-      const plainO = String(ownerDisplay).replace(/\r?\n/g, ' ').trim();
-      const lenO = plainO.length;
-      const midO = Math.floor(lenO / 2);
-      let breakO = -1;
-      const commaIdx = plainO.indexOf(',');
-      if (commaIdx !== -1 && commaIdx < lenO - 1) breakO = commaIdx + 1;
-      if (breakO === -1) {
-        const window = 24;
-        const after = plainO.slice(midO, Math.min(lenO, midO + window)).search(/[\u00A0\s]/);
-        const before = plainO.slice(Math.max(0, midO - window), midO).lastIndexOf(' ');
-        const beforeNb = plainO.slice(Math.max(0, midO - window), midO).lastIndexOf('\u00A0');
-        const bestBefore = Math.max(before, beforeNb);
-        if (after !== -1) breakO = midO + after + 1; else if (bestBefore !== -1) breakO = Math.max(1, Math.max(0, midO - window) + bestBefore + 1);
-      }
-      if (breakO <= 0 || breakO >= lenO) breakO = Math.max(1, midO);
-      const p1 = plainO.slice(0, breakO).trimEnd();
-      const p2 = plainO.slice(breakO).trimStart();
-      ownerHtml = `${escapeHtml(p1).replace(/\s/g, '\u00A0')}<br/>${escapeHtml(p2).replace(/\s/g, '\u00A0')}`;
-    } catch (_) { /* keep fallback ownerEsc */ }
-    return `<tr><td style="white-space: normal; word-break: normal; overflow-wrap: break-word; hyphens: none;"><div style="line-height:1.35; min-height: calc(2 * 1.35em); max-height: calc(2 * 1.35em); overflow: auto; white-space: nowrap;">${jobHtml}</div></td><td style="white-space: normal; word-break: normal; overflow-wrap: break-word; hyphens: none;">${testEsc}</td><td style="white-space: nowrap; word-break: keep-all; overflow: hidden;"><div style="line-height:1.35; min-height: calc(2 * 1.35em); max-height: calc(2 * 1.35em); overflow: auto; white-space: nowrap;">${ownerHtml}</div></td><td style="white-space: normal; word-break: normal; overflow-wrap: break-word; hyphens: none;">${snippetOneLine}</td></tr>`;
-  }).join('\n');
-  // Compute dynamic width for Job column based on longest job label; cap to reasonable bounds
-  const jobWidthCh = (() => {
-    const est = Math.ceil((__maxJobLen || 0) / 2) + 2; // two-line target + padding
-    const minCh = 28; // lower bound so it doesn't get too narrow
-    const maxCh = 88; // upper bound to avoid crowding other columns
-    return Math.max(minCh, Math.min(maxCh, est));
-  })();
-  // ========================================================
-
-  return [
-    '<table style="table-layout: auto; width: 100%;">',
-    '<thead>',
-    '<tr>',
-    `<th style="width: ${jobWidthCh}ch;">Job</th>`,
-    '<th>Test</th>',
-    '<th>Owner</th>',
-    '<th>Error</th>',
-    '</tr>',
-    '</thead>',
-    '<tbody>',
-    rows,
-    '</tbody>',
-    '</table>',
-    ''
-  ].join('\n');
-}
-
-// Helper: derive a one-line description for a commit
-function getCommitDescription(commit) {
-  try {
-    if (!commit) return '';
-    const fromField = (typeof commit.description === 'string' && commit.description) || '';
-    if (fromField) return fromField;
-    const msg = typeof commit.message === 'string' ? commit.message : '';
-    if (!msg) return '';
-    const firstLine = String(msg).split(/\r?\n/)[0] || '';
-    return firstLine;
-  } catch (_) {
-    return '';
-  }
-}
-
-function renderCommitsTable(commits) {
-  if (!Array.isArray(commits) || commits.length === 0) {
-    return '<p><em>None</em></p>';
-  }
-  const escapeHtml = (str) => String(str).replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;').replace(/'/g, '&#039;');
-  const rows = commits.map(c => {
-    const short = c.short || (c.sha ? c.sha.substring(0, 7) : '');
-    const url = c.url || (c.sha ? `https://github.com/${github.context.repo.owner}/${github.context.repo.repo}/commit/${c.sha}` : undefined);
-    const who = c.author_login ? `@${c.author_login}` : (c.author_name || 'unknown');
-    const whoHtml = (c.author_login && c.author_url)
-      ? `<a href="${escapeHtml(c.author_url)}">@${escapeHtml(c.author_login)}</a>`
-      : escapeHtml(who);
-    const shaHtml = url ? `<a href="${escapeHtml(url)}"><code>${escapeHtml(short)}</code></a>` : `<code>${escapeHtml(short)}</code>`;
-    const descHtml = escapeHtml(getCommitDescription(c));
-    return `<tr><td>${shaHtml}</td><td>${whoHtml}</td><td>${descHtml}</td></tr>`;
-  }).join('\n');
-  return [
-    '<table>',
-    '<thead>',
-    '<tr><th>SHA</th><th>Author</th><th>Description</th></tr>',
-    '</thead>',
-    '<tbody>',
-    rows,
-    '</tbody>',
-    '</table>',
-    ''
-  ].join('\n');
-}
-
-/**
- * Fetches PR information associated with a commit.
- *
- * @param {object} context - GitHub Actions context
- * @param {string} commitSha - Full SHA of the commit to look up
- * @returns {Promise<object>} Object containing:
- *   - prNumber: Markdown link to the PR (e.g., [#123](url))
- *   - prTitle: Title of the PR or EMPTY_VALUE if not found
- *   - prAuthor: GitHub username of the PR author or 'unknown'
- */
-// Disabled: PR fetching via GitHub API removed for offline analysis
-async function fetchPRInfo(_github, _context, _commitSha) {
-  return { prNumber: EMPTY_VALUE, prTitle: EMPTY_VALUE, prAuthor: EMPTY_VALUE };
-}
-
-/**
- * Fetch commit author info for a commit SHA.
- * Returns GitHub login (if associated), author display name, and profile URL (if available).
- */
-// Disabled: commit author fetch via GitHub API; will be inferred from commits index if present
-async function fetchCommitAuthor(_commitSha) {
-  return { login: undefined, name: undefined, htmlUrl: undefined };
-}
-
-/**
- * Download workflow run logs and extract up to N error snippets.
- * Returns an array of strings (snippets).
- */
-async function fetchErrorSnippetsForRun(runId, maxSnippets = 50, logsDirPath = undefined, annotationsDirPath = undefined) {
-  try {
-    await core.startGroup(`Extracting error snippets for run ${runId}`); // start a group to log the error snippets (we log this to the console)
-    let snippets = [];
-
-    // Helper: filter out generic exit-code snippets when a job has more specific errors
-    const isGenericExit = (s) => typeof s === 'string' && /^Process completed with exit code 1\.?$/i.test(s.trim()); // this usually has no meaningful information because a previous error that was already registered just caused this one to appear
-    const filterGenericExitSnippets = (arr) => {
-      try {
-        const byJob = new Map();
-        for (const sn of (arr || [])) {
-          const job = (sn && sn.job) ? String(sn.job) : '';
-          if (!byJob.has(job)) byJob.set(job, []);
-          byJob.get(job).push(sn);
-        }
-        const out = [];
-        for (const [job, list] of byJob.entries()) {
-          const hasNonGeneric = list.some(sn => !isGenericExit(sn && sn.snippet)); // test to see if there is a non-generic exit snippet
-          if (hasNonGeneric) {
-            for (const sn of list) { if (!isGenericExit(sn && sn.snippet)) out.push(sn); } // if there is a non-generic exit snippet, add it to the output
-          } else {
-            out.push(...list); // if there is no non-generic exit snippet, add all the snippets to the output (no point in filtering)
-          }
-        }
-        return out;
-      } catch (_) { return arr || []; }
-    };
-
-    // If gtest logs are available for this run, parse them and return. This leaves pytest/others unchanged.
-    try {
-      const runLogsDir = logsDirPath || getGtestLogsDirForRunId(runId);
-      if (runLogsDir && fs.existsSync(runLogsDir)) {
-        const idxPath = path.join(runLogsDir, 'jobs.json');
-        const extractDir = path.join(runLogsDir, 'extract');
-        if (fs.existsSync(idxPath) && fs.existsSync(extractDir)) {
-          core.info(`[GTEST] Using logs for run ${runId}: runLogsDir=${runLogsDir}`);
-          core.info(`[GTEST] Index path: ${idxPath}`);
-          let idx;
-          try {
-            idx = JSON.parse(fs.readFileSync(idxPath, 'utf8'));
-          } catch (parseErr) {
-            core.warning(`[GTEST] Failed to parse jobs.json for run ${runId}: ${parseErr.message}`);
-            // Continue to other log sources
-          }
-          if (idx) {
-            const jobs = Array.isArray(idx.jobs) ? idx.jobs : [];
-            core.info(`[GTEST] Jobs detected: ${jobs.length}`);
-            const out = [];
-            // Helper to strip ANSI color codes and other escape sequences
-            const stripAnsi = (s) => typeof s === 'string' ? s.replace(/\x1b\[[0-9;]*m/g, '') : s; // this is probably unnecessary but no harm in doing it
-
-            for (const job of jobs) {
-              const jobName = (job && job.name) ? String(job.name) : 'gtest';
-              const files = Array.isArray(job.files) ? job.files : [];
-              core.info(`[GTEST] Processing job: name='${jobName}', files=${files.length}`);
-              let lastSeenTestName = undefined; // persist across files in this job
-              for (const rel of files) {
-                if (out.length >= maxSnippets) break;
-                try {
-                  const abs = path.join(runLogsDir, rel);
-                  if (!fs.existsSync(abs)) { core.info(`[GTEST]   Skip missing file: ${abs}`); continue; }
-                  const text = fs.readFileSync(abs, 'utf8');
-                  const rawLines = text.split(/\r?\n/);
-                  const lines = rawLines.map(l =>
-                    stripAnsi(l)
-                      .replace(/^\s*\d{4}-\d{2}-\d{2}T[0-9:.]+Z\s+/, '')
-                      .replace(/^\s*\[[0-9]+,[0-9]+\]<[^>]+>:\s*/, '')
-                  );
-                  core.info(`[GTEST]   Parsing file: ${abs} (lines=${lines.length})`);
-                  let currentTest = lastSeenTestName;
-                  let capturing = false;
-                  let buf = [];
-                  let snippetsAdded = 0;
-
-                  const flush = (lineNo, iIndex) => {
-                    if (!capturing) return;
-                    const msg = buf.join('\n').trim();
-                    if (msg) {
-                      // If name unknown, try to infer by scanning forward a small window
-                      let labelName = (currentTest && String(currentTest).trim()) || '';
-                      if (!labelName) {
-                        for (let j = iIndex + 1; j < Math.min(lines.length, iIndex + 200); j++) {
-                          const l2 = lines[j];
-                          const mFail = l2 && l2.match(/\[\s*FAILED\s*\]\s+(.+?)\s*$/);
-                          if (mFail) { labelName = mFail[1]; break; }
-                          const mRun2 = l2 && l2.match(/\[\s*RUN\s*\]\s+(.+?)\s*$/);
-                          if (mRun2) { labelName = mRun2[1]; break; }
-                        }
-                      }
-                      if (!labelName) labelName = lastSeenTestName || 'unknown gtest';
-                      // Do not assign owners here; only return raw snippet with inferred job/test via label for hint
-                      out.push({ label: `${jobName}: ${labelName}`, job: jobName, test: labelName, snippet: msg });
-                      snippetsAdded++;
-                      core.info(`[GTEST]     Added snippet for '${labelName}' (len=${msg.length}) @ line ${lineNo}`);
-                    }
-                    capturing = false; buf = [];
-                  };
-
-                  for (let i = 0; i < lines.length && out.length < maxSnippets; i++) {
-                    const rawLine = lines[i];
-                    const line = stripAnsi(rawLine);
-                    // Detect new test start
-                    // Match RUN anywhere in the line (timestamps/prefixes may precede it)
-                    const runMatch = line && line.match(/\[\s*RUN\s*\]\s+(.+?)\s*$/);
-                    if (runMatch) {
-                      // Starting a new test block; stop any capture in progress (without emitting)
-                      capturing = false; buf = [];
-                      currentTest = runMatch[1];
-                      lastSeenTestName = currentTest;
-                      core.info(`[GTEST]     RUN -> '${currentTest}' @ line ${i+1}`);
-                      continue;
-                    }
-
-                    // Look for info/backtrace block boundaries
-                    const infoMatch = line && line.match(/^\s*info:\s*(.*)$/i);
-                    if (!capturing && infoMatch) {
-                      const lower = line.toLowerCase();
-                      const btIdx = lower.indexOf('backtrace:');
-                      if (btIdx !== -1) {
-                        // Same-line info..backtrace
-                        const infoIdx = lower.indexOf('info:');
-                        const between = line.substring(infoIdx + 5, btIdx).replace(/^\s*|\s*$/g, '');
-                        if (between) buf.push(between);
-                        flush(i+1, i);
-                        continue;
-                      }
-                      // Start multi-line capture
-                      capturing = true;
-                      buf = [];
-                      if (infoMatch[1]) buf.push(infoMatch[1]);
-                      core.info(`[GTEST]       info: begin capture @ line ${i+1}`);
-                      continue;
-                    }
-
-                    if (capturing) {
-                      const lower = line.toLowerCase();
-                      const btIdx2 = lower.indexOf('backtrace:');
-                      if (btIdx2 !== -1) {
-                        const head = line.substring(0, btIdx2).replace(/^\s*|\s*$/g, '');
-                        if (head) buf.push(head);
-                        flush(i+1, i);
-                        continue;
-                      }
-                      buf.push(line);
-                    }
-                  }
-                  core.info(`[GTEST]   File complete: snippets_added=${snippetsAdded}`);
-                } catch (errFile) {
-                  core.info(`[GTEST]   Failed parsing file '${rel}': ${errFile && errFile.message || String(errFile)}`);
-                }
-              }
-            }
-            if (out.length) {
-              // Dedupe by normalized label+snippet to avoid repeated rows
-              const norm = (s) => String(s || '').replace(/\s+/g, ' ').trim();
-              const seen = new Set();
-              const unique = [];
-              for (const e of out) {
-                const key = `${norm(e && e.label)}|${norm(e && e.snippet)}`;
-                if (!seen.has(key)) { seen.add(key); unique.push(e); }
-              }
-              core.info(`[GTEST] Total snippets before cap: ${out.length}, after dedupe: ${unique.length}`);
-              snippets = unique.slice(0, Math.max(1, Math.min(maxSnippets, unique.length)));
-              // Apply generic-exit filter per job
-              snippets = filterGenericExitSnippets(snippets);
-              core.info(`[GTEST] Collected ${snippets.length} gtest snippet(s) from logs for run ${runId} (after filtering)`);
-              // Do not attach owners here; ownership resolution is performed later
-              return snippets;
-            } else {
-              core.info(`[GTEST] No snippets extracted from logs for run ${runId}`);
-            }
-          }
-        }
-      }
-    } catch (e) {
-      core.warning(`Failed gtest log parsing for run ${runId}: ${e.message}`);
-    }
-
-    // If other logs (non-gtest) are available for this run, extract job names from file names
-    // Don't parse the logs, just use the file names to determine job names
-    try {
-      const otherLogsDir = getOtherLogsDirForRunId(runId);
-      if (otherLogsDir && fs.existsSync(otherLogsDir)) {
-        const logsListPath = path.join(otherLogsDir, 'logs-list.json');
-        if (fs.existsSync(logsListPath)) {
-          core.info(`[OTHER LOGS] Using logs list for run ${runId}: otherLogsDir=${otherLogsDir}`);
-          let logsListData;
-          try {
-            logsListData = JSON.parse(fs.readFileSync(logsListPath, 'utf8'));
-          } catch (parseErr) {
-            core.warning(`[OTHER LOGS] Failed to parse logs-list.json for run ${runId}: ${parseErr.message}`);
-            // Continue to annotations
-          }
-          if (logsListData) {
-            const files = Array.isArray(logsListData.files) ? logsListData.files : [];
-            core.info(`[OTHER LOGS] Files detected: ${files.length}`);
-
-            // Extract job names from file paths (e.g., "extract/1_job-name/step.txt" -> "job-name")
-            const jobNamesSet = new Set();
-            for (const filePath of files) {
-              try {
-                // Parse the path to extract job name
-                // Expected format: extract/<step>_<job-name>/<file>.txt
-                const parts = filePath.split(path.sep);
-                if (parts.length >= 2) {
-                  const folderName = parts[1]; // e.g., "1_job-name"
-                  // Remove leading step number and underscore
-                  const jobName = folderName.replace(/^\d+_/, '').trim();
-                  if (jobName) {
-                    jobNamesSet.add(jobName);
-                  }
-                }
-              } catch (_) { /* ignore */ }
-            }
-
-            // Create snippets with job names, but blank test and informative error message
-            const out = [];
-            for (const jobName of jobNamesSet) {
-              out.push({
-                label: jobName,
-                job: jobName,
-                test: '',
-                snippet: 'currently aggregate-workflow-data is not able to parse these kinds of errors'
-              });
-              if (out.length >= maxSnippets) break;
-            }
-
-            if (out.length > 0) {
-              core.info(`[OTHER LOGS] Collected ${out.length} job name(s) from logs for run ${runId}`);
-              snippets = out;
-              // Apply generic-exit filter per job
-              snippets = filterGenericExitSnippets(snippets);
-              return snippets;
-            } else {
-              core.info(`[OTHER LOGS] No job names extracted from logs for run ${runId}`);
-            }
-          }
-        }
-      }
-    } catch (e) {
-      core.warning(`Failed other log processing for run ${runId}: ${e.message}`);
-    }
-
-    // Primary (default): use annotations if available (non-gtest)
-    // If we already collected snippets from logs, skip annotations entirely
-    if (Array.isArray(snippets) && snippets.length > 0) {
-      return snippets;
-    }
-    const annDir = annotationsDirPath || getAnnotationsDirForRunId(runId);
-    if (annDir && fs.existsSync(annDir)) {
-      try {
-        const filePath = path.join(annDir, 'annotations.json');
-        if (fs.existsSync(filePath)) {
-          const raw = fs.readFileSync(filePath, 'utf8');
-          const arr = JSON.parse(raw);
-          if (Array.isArray(arr)) {
-            const seen = new Set();
-            for (const a of arr) {
-              const job = a.job_name || '';
-              const title = a.title || '';
-              const level = a.annotation_level || '';
-              const message = a.message || '';
-              const levelLc = String(level).toLowerCase();
-              if (levelLc !== 'failure' && levelLc !== 'error') continue; // ignore warnings
-              const msgTrim = String(message).trim();
-              const label = `${job}${title ? `: ${title}` : ''} [${level}]`;
-              const dedupeKey = `${job}|${title}|${levelLc}|${msgTrim}`;
-              if (seen.has(dedupeKey)) continue;
-              seen.add(dedupeKey);
-              snippets.push({ label, job, snippet: message });
-              if (snippets.length >= maxSnippets) break;
-            }
-          }
-        }
-      } catch (e) {
-        core.warning(`Failed reading annotations for run ${runId}: ${e.message}`);
-      }
-    }
-    // Fallback: none (we no longer parse logs). If needed, could fallback to logs here.
-    // Do not attach owners here; ownership resolution is performed later
-
-    core.info(`Total snippets collected from annotations: ${snippets.length}`);
-    // Apply generic-exit filter per job before returning
-    return filterGenericExitSnippets(snippets || []);
-  } catch (e) {
-    core.info(`Failed to obtain run logs for ${runId}: ${e.message}`);
-    return [];
-  } finally {
-    core.endGroup();
-  }
-}
-
-// END OF CODE FOR ERROR HANDLING
-
-/**
- * Calculates statistics for a set of workflow runs.
- *
- * @param {Array<object>} runs - Array of workflow run objects
- * @returns {object} Statistics object containing:
- *   - eventTypes: Comma-separated string of unique event types
- *   - successRate: Percentage of runs that succeeded (based on last attempt)
- *   - uniqueSuccessRate: Percentage of runs that succeeded on first attempt
- *   - retryRate: Percentage of successful runs that required retries
- *   - uniqueRuns: Number of unique runs (excluding attempts)
- *   - totalRuns: Total number of runs including attempts
- *   - successfulRuns: Number of runs that succeeded (based on last attempt)
- */
-function getWorkflowStats(runs) {
-  // Group runs by their original run ID to handle retries and reruns
-  const uniqueRuns = new Map();
-  let totalRunsIncludingRetries = 0;
-  let totalSuccessfulUniqueRuns = 0;
-  let successfulUniqueRunsOnFirstTry = 0;
-  let successfulUniqueRunsWithRetries = 0;
-
-  // First pass: identify all unique runs and their attempts
-  for (const run of runs) {
-    totalRunsIncludingRetries++;
-
-    // GitHub keeps the same run ID for re-runs, only incrementing run_attempt
-    // So we group by run.id directly, not by calculating a different ID
-    const runId = run.id;
-    const currentAttempt = run.run_attempt || 1;
-
-    if (!uniqueRuns.has(runId)) {
-      // First time seeing this run ID
-      uniqueRuns.set(runId, {
-        run,
-        attempts: 1, // Initialize to 1 since this is the first attempt
-        isSuccessful: false,
-        requiredRetry: false,
-        succeededOnFirstTry: false,
-        lastAttempt: run
-      });
-    } else {
-      // This is a re-run (same run ID, different attempt)
-      const existingRun = uniqueRuns.get(runId);
-      existingRun.attempts++; // increment the attempts counter
-
-      // Update last attempt if this is a newer attempt (higher run_attempt number)
-      const existingAttempt = existingRun.lastAttempt.run_attempt || 1;
-      if (currentAttempt > existingAttempt) {
-        existingRun.lastAttempt = run; // update to the latest attempt
-      }
-    }
-  }
-
-  // Second pass: determine final status based on last attempt
-  for (const runInfo of uniqueRuns.values()) { // iterate through the unique runs
-    const lastAttempt = runInfo.lastAttempt;
-    if (lastAttempt.conclusion === 'success') { // if the last attempt is successful
-      runInfo.isSuccessful = true;
-      totalSuccessfulUniqueRuns++; // increment the total successful unique runs
-
-      if (lastAttempt.run_attempt === 1) { // if the last attempt is the first attempt
-        runInfo.succeededOnFirstTry = true; // set the succeeded on first try to true
-        successfulUniqueRunsOnFirstTry++; // increment the total successful unique runs on first try
-      } else {
-        runInfo.requiredRetry = true; // set the required retry to true
-        successfulUniqueRunsWithRetries++; // increment the total successful unique runs with retries
-      }
-    }
-  }
-
-  const uniqueRunsArray = Array.from(uniqueRuns.values()).map(r => r.run); // creates an array of the run objects
-  const eventTypes = [...new Set(uniqueRunsArray.map(r => r.event))].join(', '); // create a string of all the even types that were used to trigger runs
-
-  // Calculate rates
-  const successRate = uniqueRunsArray.length === 0 ? "N/A" : (totalSuccessfulUniqueRuns / uniqueRunsArray.length * 100).toFixed(SUCCESS_RATE_DECIMAL_PLACES) + "%";
-  const uniqueSuccessRate = uniqueRunsArray.length === 0 ? "N/A" : (successfulUniqueRunsOnFirstTry / uniqueRunsArray.length * 100).toFixed(SUCCESS_RATE_DECIMAL_PLACES) + "%";
-  const retryRate = totalSuccessfulUniqueRuns === 0 ? "N/A" : (successfulUniqueRunsWithRetries / totalSuccessfulUniqueRuns * 100).toFixed(SUCCESS_RATE_DECIMAL_PLACES) + "%";
-
-  return { // return the eventypes, success rates, retry rates, and total run info
-    eventTypes,
-    successRate,
-    uniqueSuccessRate,
-    retryRate,
-    uniqueRuns: uniqueRunsArray.length,
-    totalRuns: totalRunsIncludingRetries,
-    successfulRuns: totalSuccessfulUniqueRuns
-  };
-}
-
-/**
- * Generates a GitHub Actions workflow URL.
- *
- * @param {object} context - GitHub Actions context
- * @param {string} workflowFile - Path to the workflow file relative to .github/workflows/
- * @returns {string} Full URL to the workflow in GitHub Actions
- */
-function getWorkflowLink(context, workflowFile) {
-  return workflowFile // return the workflow link querying on main branch if the workflow file exists
-    ? `https://github.com/${context.repo.owner}/${context.repo.repo}/actions/workflows/${workflowFile}?query=branch%3Amain`
-    : `https://github.com/${context.repo.owner}/${context.repo.repo}/actions`; // return the github actions link if the workflow file does not exist
-}
-
-/**
- * Finds, within the provided window (newest→oldest, main branch only), either:
- * - the first failing run since the most recent success (oldest in the current failing streak), or
- * - if no success exists in-window, the oldest failing run in the window.
- * Returns null if there are no failing runs in the window.
- *
- * @param {Array<object>} mainBranchRunsWindow - Runs on main, sorted by created_at desc (newest first)
- * @returns {{run: object, noSuccessInWindow: boolean}|null}
- */
-function findFirstFailInWindow(mainBranchRunsWindow) {
-  let seenAnyFailure = false;
-  let firstFailInStreak = null; // oldest failure observed before crossing a success boundary
-
-  for (const run of mainBranchRunsWindow) {
-    if (run.conclusion === 'success') {
-      if (firstFailInStreak) {
-        // We found a success after observing failures: return the oldest failure in the streak
-        return { run: firstFailInStreak, boundarySuccessRun: run, noSuccessInWindow: false };
-      }
-      // Success encountered before any failure in the current scan; keep scanning older entries
-    } else if (run.conclusion && run.conclusion !== 'cancelled' && run.conclusion !== 'skipped') {
-      // Treat anything non-success, non-cancelled/skipped as failure for this purpose
-      seenAnyFailure = true;
-      firstFailInStreak = run; // update to become oldest failure within the current failing streak (and window)
-    }
-  }
-
-  if (seenAnyFailure) {
-    // No success found in-window; report oldest failure in the window
-    return { run: firstFailInStreak, boundarySuccessRun: undefined, noSuccessInWindow: true };
-  }
-  return null;
-}
-
-/**
- * Finds the first failing run on main since the last success (i.e., the start of the current failing streak).
- * Scans runs in reverse chronological order and returns the oldest failure before the first encountered success.
- * Falls back to the oldest failure in history if no success is found.
- *
- * @param {object} context - GitHub Actions context
- * @param {string} workflowPath - Path to the workflow file (e.g., .github/workflows/ci.yaml)
- * @returns {Promise<object|null>} The workflow run object or null if none found
- */
-
-/**
- * Analyzes scheduled runs to find the last good and earliest bad commits.
- *
- * @param {Array<object>} scheduledMainRuns - Array of scheduled runs on main branch, sorted by date (newest first)
- * @param {object} context - GitHub Actions context
- * @returns {object} Object containing:
- *   - newestGoodSha: Short SHA of the most recent successful run (e.g., `a1b2c3d`)
- *   - newestBadSha: Short SHA of the most recent failing run (e.g., `e4f5g6h`)
- */
-function findGoodBadCommits(scheduledMainRuns, context) {
-  let newestGoodSha = EMPTY_VALUE;
-  let newestBadSha = EMPTY_VALUE;
-  let foundGood = false;
-  let foundBad = false;
-
-  for (const run of scheduledMainRuns) {
-    if (!foundGood && run.conclusion === 'success') { // find the most recent successful run
-      const shortSha = run.head_sha.substring(0, SHA_SHORT_LENGTH);
-      newestGoodSha = `[\`${shortSha}\`](https://github.com/${context.repo.owner}/${context.repo.repo}/commit/${run.head_sha})`; // set the newest good sha to the short sha of the most recent successful run
-      foundGood = true;
-    }
-    if (!foundBad && run.conclusion !== 'success') { // find the most recent failing run
-      const shortSha = run.head_sha.substring(0, SHA_SHORT_LENGTH);
-      newestBadSha = `[\`${shortSha}\`](https://github.com/${context.repo.owner}/${context.repo.repo}/commit/${run.head_sha})`; // set the newest bad sha to the short sha of the most recent failing run
-      foundBad = true;
-    }
-    if (foundGood && foundBad) break; // break the loop if both good and bad runs are found
-  }
-
-  return { newestGoodSha, newestBadSha };
-}
-
-/**
- * Gets information about the last run on main branch.
- *
- * @param {Array<object>} mainBranchRuns - Array of runs on main branch, sorted by date (newest first)
- * @param {object} context - GitHub Actions context
- * @returns {Promise<object>} Object containing run information
- */
-async function getLastRunInfo(mainBranchRuns, context) {
-  const lastMainRun = mainBranchRuns[0];
-  if (!lastMainRun) { // if there is no last main run, return the empty values
-    return {
-      status: EMPTY_VALUE,
-      sha: EMPTY_VALUE,
-      run: EMPTY_VALUE,
-      pr: EMPTY_VALUE,
-      title: EMPTY_VALUE,
-      newestGoodSha: EMPTY_VALUE,
-      newestBadSha: EMPTY_VALUE
-    };
-  }
-
-  const prInfo = await fetchPRInfo(null, context, lastMainRun.head_sha); // fetch the PR info for the last main run
-  // Current approach: filter by event type
-  const mainRuns = mainBranchRuns.filter(r => r.event === lastMainRun.event || r.event === 'workflow_dispatch'); // get the relevant main runs for finding good and bad commits
-  // Alternative approach: include all runs on main branch
-  // const mainRuns = mainBranchRuns;
-  const { newestGoodSha, newestBadSha } = findGoodBadCommits(mainRuns, context); // find the newest good and bad commits
-
-  return {
-    status: lastMainRun.conclusion === 'success' ? SUCCESS_EMOJI : FAILURE_EMOJI,
-    sha: `[\`${lastMainRun.head_sha.substring(0, SHA_SHORT_LENGTH)}\`](https://github.com/${context.repo.owner}/${context.repo.repo}/commit/${lastMainRun.head_sha})`, // get the short sha hyperlink for the latest main run
-    run: `[Run](${lastMainRun.html_url})${lastMainRun.run_attempt > 1 ? ` (#${lastMainRun.run_attempt})` : ''}`, // get the run hyperlink for the latest main run
-    pr: prInfo.prNumber, // get the PR number for the latest main run
-    title: prInfo.prTitle, // get the PR title for the latest main run
-    newestGoodSha, // get the newest good sha for the latest main run
-    newestBadSha: lastMainRun.conclusion !== 'success' ? newestBadSha : EMPTY_VALUE // if the last main run is successful, don't bother showing the most recent failure
-  };
-}
-
-/**
- * Generates summary tables for push and scheduled workflows.
- *
- * @param {Map<string, Array<object>>} grouped - Map of workflow names to their runs
- * @param {object} context - GitHub Actions context
- * @returns {Promise<string>} Markdown table for all workflows
- */
-async function generateSummaryBox(grouped, context) {
-  const rows = [];
-  const escapeHtml = (str) => String(str).replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;').replace(/'/g, '&#039;');
-
-  // Helper to convert markdown links to HTML
-  const mdToHtml = (md) => {
-    if (!md || md === EMPTY_VALUE) return escapeHtml(md);
-    // Match markdown link format: [text](url) or [`text`](url)
-    return md.replace(/\[`([^`]+)`\]\(([^)]+)\)/g, (_, text, url) => `<a href="${escapeHtml(url)}"><code>${escapeHtml(text)}</code></a>`)
-             .replace(/\[([^\]]+)\]\(([^)]+)\)/g, (_, text, url) => `<a href="${escapeHtml(url)}">${escapeHtml(text)}</a>`);
-  };
-
-  for (const [name, runs] of grouped.entries()) {
-    // First deduplicate by run ID, keeping highest attempt
-    const runsByID = new Map();
-    for (const run of runs.filter(r => r.head_branch === 'main')) {
-      const runId = run.id;
-      const currentAttempt = run.run_attempt || 1;
-      const existingRun = runsByID.get(runId);
-      const existingAttempt = existingRun ? (existingRun.run_attempt || 1) : 0;
-      if (!existingRun || currentAttempt > existingAttempt) {
-        runsByID.set(runId, run);
-      }
-    }
-    const mainBranchRuns = Array.from(runsByID.values())
-      .sort((a, b) => {
-        // Sort by date (newest first), then by run_attempt (highest first) as tiebreaker
-        const dateDiff = new Date(b.created_at) - new Date(a.created_at);
-        if (dateDiff !== 0) {
-          return dateDiff;
-        }
-        const attemptA = a.run_attempt || 1;
-        const attemptB = b.run_attempt || 1;
-        return attemptB - attemptA; // Prefer higher attempt number
-      });
-
-    const stats = getWorkflowStats(runs); // get the workflow stats for the workflow
-
-    // Skip workflows that only have workflow_dispatch as their event type. Presumably because if they have to be triggered to run they aren't super important
-    if (stats.eventTypes === 'workflow_dispatch') {
-      continue; // if the workflow only has workflow_dispatch as its event type, skip it
-    }
-
-    const workflowLink = getWorkflowLink(context, runs[0]?.path); // get the link to the workflow page that lists all the runs
-    const runInfo = await getLastRunInfo(mainBranchRuns, context); // get the last run info for the workflow
-
-    // basically, create the statistics table for the workflow. these are the columns that are displayed in the summary box.
-    const row = `<tr>
-<td><a href="${escapeHtml(workflowLink)}">${escapeHtml(name)}</a></td>
-<td>${escapeHtml(stats.eventTypes || 'unknown')}</td>
-<td>${stats.totalRuns}</td>
-<td>${stats.successfulRuns}</td>
-<td>${escapeHtml(stats.successRate)}</td>
-<td>${escapeHtml(stats.uniqueSuccessRate)}</td>
-<td>${escapeHtml(stats.retryRate)}</td>
-<td>${escapeHtml(runInfo.status)}</td>
-<td>${mdToHtml(runInfo.sha)}</td>
-<td>${mdToHtml(runInfo.run)}</td>
-<td>${mdToHtml(runInfo.pr)}</td>
-<td>${escapeHtml(runInfo.title)}</td>
-<td>${mdToHtml(runInfo.newestBadSha)}</td>
-<td>${mdToHtml(runInfo.newestGoodSha)}</td>
-</tr>`;
-    rows.push(row);
-  }
-
-  return [
-    '## Workflow Summary',
-    '<table>',
-    '<thead>',
-    '<tr><th>Workflow</th><th>Event Type(s)</th><th>Total Runs</th><th>Successful Runs</th><th>Success Rate</th><th>Unique Success Rate</th><th>Retry Rate</th><th>Last Run on <code>main</code></th><th>Last SHA</th><th>Last Run</th><th>Last PR</th><th>PR Title</th><th>Newest Bad SHA</th><th>Newest Good SHA</th></tr>',
-    '</thead>',
-    '<tbody>',
-    ...rows,
-    '</tbody>',
-    '</table>',
-    ''
-  ].join('\n');
-}
-
-/**
- * Builds the complete markdown report.
- *
- * @param {Map<string, Array<object>>} grouped - Map of workflow names to their runs
- * @param {object} context - GitHub Actions context
- * @returns {Promise<string>} Complete markdown report
- */
-async function buildReport(grouped, context) {
-  const days = parseInt(core.getInput('days') || DEFAULT_LOOKBACK_DAYS, 10); // get the number of days to look back for workflow data
-  const timestamp = new Date().toISOString(); // get the timestamp for the report
-  return [
-    `# Workflow Summary (Last ${days} Days) - Generated at ${timestamp}\n`,
-    await generateSummaryBox(grouped, context),
-    '\n## Column Descriptions\n',
-    '<p>A unique run represents a single workflow execution, which may have multiple retry attempts. For example, if a workflow fails and is retried twice, this counts as one unique run with three attempts (initial run + two retries).</p>\n',
-    '\n### Success Rate Calculations\n',
-    '<p>The success rates are calculated based on unique runs (not including retries in the denominator):</p>\n',
-    '<ul>',
-    '<li><strong>Success Rate</strong>: (Number of unique runs that eventually succeeded / Total number of unique runs) × 100%',
-    '  <ul><li>Example: 3 successful unique runs out of 5 total unique runs = 60% success rate</li></ul>',
-    '</li>',
-    '<li><strong>Unique Success Rate</strong>: (Number of unique runs that succeeded on first try / Total number of unique runs) × 100%',
-    '  <ul><li>Example: 1 unique run succeeded on first try out of 5 total unique runs = 20% unique success rate</li></ul>',
-    '</li>',
-    '<li><strong>Retry Rate</strong>: (Number of successful unique runs that needed retries / Total number of successful unique runs) × 100%',
-    '  <ul><li>Example: 2 successful unique runs needed retries out of 3 total successful unique runs = 66.67% retry rate</li></ul>',
-    '</li>',
-    '</ul>\n',
-    '<p><strong>Note:</strong> Unique Success Rate + Retry Rate does not equal 100% because they measure different things:</p>',
-    '<ul>',
-    '<li>Unique Success Rate is based on all unique runs</li>',
-    '<li>Retry Rate is based only on successful unique runs</li>',
-    '</ul>\n',
-    '<table>',
-    '<thead>',
-    '<tr><th>Column</th><th>Description</th></tr>',
-    '</thead>',
-    '<tbody>',
-    '<tr><td>Workflow</td><td>Name of the workflow with link to its GitHub Actions page</td></tr>',
-    '<tr><td>Event Type(s)</td><td>Types of events that trigger this workflow (e.g., push, pull_request, schedule)</td></tr>',
-    '<tr><td>Total Runs</td><td>Total number of workflow runs including all retry attempts (e.g., 1 unique run with 2 retries = 3 total runs)</td></tr>',
-    '<tr><td>Successful Runs</td><td>Number of unique workflow runs that eventually succeeded, regardless of whether they needed retries</td></tr>',
-    '<tr><td>Success Rate</td><td>Percentage of unique workflow runs that eventually succeeded (e.g., 3/5 unique runs succeeded = 60%)</td></tr>',
-    '<tr><td>Unique Success Rate</td><td>Percentage of unique workflow runs that succeeded on their first attempt without needing retries (e.g., 1/5 unique runs succeeded on first try = 20%)</td></tr>',
-    '<tr><td>Retry Rate</td><td>Percentage of successful unique runs that needed retries to succeed (e.g., of 3 successful unique runs, 2 needed retries = 66.67%)</td></tr>',
-    '<tr><td>Last Run on <code>main</code></td><td>Status of the most recent run on the main branch (✅ for success, ❌ for failure)</td></tr>',
-    '<tr><td>Last SHA</td><td>Short SHA of the most recent run on main</td></tr>',
-    '<tr><td>Last Run</td><td>Link to the most recent run on main, with attempt number if applicable</td></tr>',
-    '<tr><td>Last PR</td><td>Link to the PR associated with the most recent run, if any</td></tr>',
-    '<tr><td>PR Title</td><td>Title of the PR associated with the most recent run, if any</td></tr>',
-    '<tr><td>Newest Bad SHA</td><td>Short SHA of the most recent failing run on main (only shown if last run failed)</td></tr>',
-    '<tr><td>Newest Good SHA</td><td>Short SHA of the most recent successful run on main</td></tr>',
-    '</tbody>',
-    '</table>'
-  ].join('\n');
-}
-
-/**
- * Filters workflow runs by date range
- * @param {Array<Object>} runs - Array of workflow runs
- * @param {number} days - Number of days to look back
- * @returns {Array<Object>} Filtered runs within the date range
- */
-function filterRunsByDate(runs, days) {
-  const cutoffDate = new Date();
-  cutoffDate.setDate(cutoffDate.getDate() - days); // set the cutoff date to the number of days ago
-
-  return runs.filter(run => {
-    const runDate = new Date(run.created_at);
-    return runDate >= cutoffDate; // return the runs that are within the date range
-  });
-}
-
-// START MORE ERROR HANDLING CODE
-
-/**
- * Collect commits between two SHAs on the default branch (main), inclusive of endSha.
- * Returns an array of { sha, short, url, author_login, author_name, author_url }.
- * Note: Uses compareCommits, which is base..head; base is typically the success commit, head is the failed run commit.
- */
-// Offline commits lookup: read from commits index built by fetch step
-let __commitsIndex = undefined;
-function loadCommitsIndex(commitsPath) {
-  if (!commitsPath) return undefined;
-  try {
-    if (!fs.existsSync(commitsPath)) return undefined;
-    const raw = fs.readFileSync(commitsPath, 'utf8');
-    const arr = JSON.parse(raw);
-    if (Array.isArray(arr)) return arr;
-  } catch (_) { /* ignore */ }
-  return undefined;
-}
-
-function listCommitsBetweenOffline(context, startShaExclusive, endShaInclusive) { // get all the commits between two SHAs on the default branch (main), inclusive of endSha
-  if (!Array.isArray(__commitsIndex) || __commitsIndex.length === 0) return [];
-  const commits = __commitsIndex;
-  const startIdx = commits.findIndex(c => c.sha === startShaExclusive);
-  const endIdx = commits.findIndex(c => c.sha === endShaInclusive);
-  if (endIdx === -1) return [];
-  // We want commits strictly after start and up to and including end
-  const from = startIdx === -1 ? 0 : (startIdx + 1); // basically, if the start index is -1, then we start from the beginning of the commits array
-  const to = endIdx;
-  const slice = commits.slice(from, to + 1);
-  return slice.map(c => ({
-    sha: c.sha,
-    short: (c.short || (c.sha ? c.sha.substring(0, SHA_SHORT_LENGTH) : '')),
-    url: c.url || (c.sha ? `https://github.com/${context.repo.owner}/${context.repo.repo}/commit/${c.sha}` : ''),
-    author_login: c.author_login,
-    author_name: c.author_name,
-    author_url: c.author_url,
-    message: c.message,
-    description: getCommitDescription(c),
-  }));
-}
-
-// END MORE ERROR HANDLING CODE
 
 /**
  * Main function to run the action
@@ -1278,44 +108,49 @@ async function run() {
 
 
     if (annotationsIndexPath) {
-      __annotationsIndexMap = loadAnnotationsIndexFromFile(annotationsIndexPath);
-      if (__annotationsIndexMap && __annotationsIndexMap.size) {
-        core.info(`Loaded annotations index with ${__annotationsIndexMap.size} entries from ${annotationsIndexPath}`);
+      const annotationsIndexMap = loadAnnotationsIndexFromFile(annotationsIndexPath);
+      setAnnotationsIndexMap(annotationsIndexMap);
+      if (annotationsIndexMap && annotationsIndexMap.size) {
+        core.info(`Loaded annotations index with ${annotationsIndexMap.size} entries from ${annotationsIndexPath}`);
       } else if (annotationsIndexPath) {
         core.info(`No valid entries found in annotations index file at ${annotationsIndexPath}`);
       }
     }
 
     if (gtestLogsIndexPath) {
-      __gtestLogsIndexMap = loadLogsIndexFromFile(gtestLogsIndexPath);
-      if (__gtestLogsIndexMap && __gtestLogsIndexMap.size) {
-        core.info(`Loaded gtest logs index with ${__gtestLogsIndexMap.size} entries from ${gtestLogsIndexPath}`);
+      const gtestLogsIndexMap = loadLogsIndexFromFile(gtestLogsIndexPath);
+      setGtestLogsIndexMap(gtestLogsIndexMap);
+      if (gtestLogsIndexMap && gtestLogsIndexMap.size) {
+        core.info(`Loaded gtest logs index with ${gtestLogsIndexMap.size} entries from ${gtestLogsIndexPath}`);
       } else if (gtestLogsIndexPath) {
         core.info(`No valid entries found in gtest logs index file at ${gtestLogsIndexPath}`);
       }
     }
 
     if (otherLogsIndexPath) {
-      __otherLogsIndexMap = loadLogsIndexFromFile(otherLogsIndexPath);
-      if (__otherLogsIndexMap && __otherLogsIndexMap.size) {
-        core.info(`Loaded other logs index with ${__otherLogsIndexMap.size} entries from ${otherLogsIndexPath}`);
+      const otherLogsIndexMap = loadLogsIndexFromFile(otherLogsIndexPath);
+      setOtherLogsIndexMap(otherLogsIndexMap);
+      if (otherLogsIndexMap && otherLogsIndexMap.size) {
+        core.info(`Loaded other logs index with ${otherLogsIndexMap.size} entries from ${otherLogsIndexPath}`);
       } else if (otherLogsIndexPath) {
         core.info(`No valid entries found in other logs index file at ${otherLogsIndexPath}`);
       }
     }
 
     if (lastSuccessTimestampsPath) {
-      __lastSuccessTimestamps = loadLastSuccessTimestamps(lastSuccessTimestampsPath);
-      if (__lastSuccessTimestamps && __lastSuccessTimestamps.size) {
-        core.info(`Loaded last success timestamps with ${__lastSuccessTimestamps.size} entries from ${lastSuccessTimestampsPath}`);
+      const lastSuccessTimestamps = loadLastSuccessTimestamps(lastSuccessTimestampsPath);
+      setLastSuccessTimestamps(lastSuccessTimestamps);
+      if (lastSuccessTimestamps && lastSuccessTimestamps.size) {
+        core.info(`Loaded last success timestamps with ${lastSuccessTimestamps.size} entries from ${lastSuccessTimestampsPath}`);
       } else {
         core.info(`No valid entries found in last success timestamps file at ${lastSuccessTimestampsPath}`);
       }
     }
 
     // Load commits index (optional)
-    __commitsIndex = loadCommitsIndex(commitsPath) || [];
-    core.info(`Loaded commits index entries: ${Array.isArray(__commitsIndex) ? __commitsIndex.length : 0}`);
+    const commitsIndex = loadCommitsIndex(commitsPath) || [];
+    setCommitsIndex(commitsIndex);
+    core.info(`Loaded commits index entries: ${Array.isArray(commitsIndex) ? commitsIndex.length : 0}`);
 
     // Track failed workflows
     const failedWorkflows = [];
