+++ .github/actions/fetch-workflow-data/fetch-workflow-data.js
@@ -9,477 +9,34 @@ const fs = require('fs');
 const path = require('path');
 const { execFileSync } = require('child_process');
 
-// Constants for pagination and filtering
-const MAX_PAGES = 100; // Maximum number of pages to fetch from GitHub API (tune for rate limits/performance)
-const RUNS_PER_PAGE = 100; // GitHub API max per page
-const DEFAULT_DAYS = 15; // Default rolling window in days
-/**
- * Get the cutoff date for filtering runs.
- * @param {number} days - Number of days to look back
- * @returns {Date} The cutoff date
- */
-function getCutoffDate(days) {
-  const d = new Date();
-  d.setDate(d.getDate() - days);
-  return d;
-}
+// Import modules
+const fetching = require('./lib/fetching');
+const cache = require('./lib/cache');
+const logs = require('./lib/logs');
 
-/**
- * Fetch workflow runs for specific workflows, paginated, stopping early when cached runs are encountered.
- * @param {object} github - Octokit client
- * @param {object} context - GitHub Actions context
- * @param {number} days - Number of days to look back
- * @param {Set<string>} cachedRunIds - Set of run IDs that are already cached (skip these)
- * @param {Array<string>} workflowIds - Array of workflow file paths to fetch runs for
- * @param {string} eventType - Optional event type filter
- * @returns {Promise<Array>} Array of workflow run objects (only new, non-cached runs)
- */
-async function fetchAllWorkflowRuns(github, context, days, cachedRunIds = null, workflowIds = null, eventType='') {
-  const allRuns = [];
-  const cutoffDate = getCutoffDate(days);
-  const twoWeeksAgo = getCutoffDate(14); // For early exit check
-  const cachedIds = cachedRunIds || new Set();
+// Import constants and functions used in this file
+const {
+  DEFAULT_DAYS,
+  getCutoffDate,
+  fetchAllWorkflowRuns,
+  groupRunsByName,
+} = fetching;
 
-  core.info(`[FETCH] cutoffDate: ${cutoffDate.toISOString()}`);
-  core.info(`[FETCH] days: ${days}, cachedRunIds: ${cachedIds.size}, eventType: ${eventType || 'all'}`);
-  core.info(`[FETCH] workflowIds: ${workflowIds ? workflowIds.length + ' workflows' : 'all workflows (backward compatibility)'}`);
+const {
+  findPreviousAggregateRun,
+  pruneOldRuns,
+  pruneAnnotationsIndex,
+  pruneLogsIndex,
+  pruneCommits,
+  downloadAndExtractArtifact,
+  findFileInDirectory,
+  removeUnkeptDirectories,
+} = cache;
 
-  const MAX_CONSECUTIVE_CACHED = 50; // Stop after 50 consecutive cached runs
-  let totalSkippedOldRuns = 0;
-  let totalSkippedCachedRuns = 0;
-  let totalAddedNewRuns = 0;
+const {
+  workflowMatchesConfig,
+} = logs;
 
-  // If workflow IDs are provided, fetch runs for each workflow specifically
-  if (workflowIds && workflowIds.length > 0) {
-    core.info(`[FETCH] Fetching runs for ${workflowIds.length} specific workflows`);
-    const delay = ms => new Promise(resolve => setTimeout(resolve, ms));
-
-    for (const workflowId of workflowIds) {
-      let consecutiveCachedRuns = 0;
-      let skippedOldRuns = 0;
-      let skippedCachedRuns = 0;
-      let addedNewRuns = 0;
-      let shouldStopWorkflow = false;
-
-      core.info(`[FETCH] Fetching runs for workflow: ${workflowId}`);
-
-      for (let page = 1; page <= MAX_PAGES; page++) {
-        core.info(`[FETCH] ${workflowId} - Fetching page ${page}...`);
-        const params = {
-          owner: context.repo.owner,
-          repo: context.repo.repo,
-          workflow_id: workflowId,
-          per_page: RUNS_PER_PAGE,
-          page,
-        };
-        if (eventType) {
-          params.event = eventType;
-        }
-
-        try {
-          const { data: runs } = await github.rest.actions.listWorkflowRuns(params);
-          if (!runs.workflow_runs || runs.workflow_runs.length === 0) {
-            core.info(`[FETCH] ${workflowId} - No runs on page ${page}, stopping`);
-            break;
-          }
-
-          core.info(`[FETCH] ${workflowId} - Page ${page}: processing ${runs.workflow_runs.length} runs`);
-
-          for (const run of runs.workflow_runs) {
-            const runDate = new Date(run.created_at);
-            const runIdStr = String(run.id);
-
-            // Early exit if runs are older than 2 weeks
-            if (runDate < twoWeeksAgo) {
-              core.info(`[FETCH] ${workflowId} - Early exit: runs older than 2 weeks (${runDate.toISOString()} < ${twoWeeksAgo.toISOString()})`);
-              shouldStopWorkflow = true;
-              break;
-            }
-
-            // Skip runs older than cutoff date
-            if (runDate < cutoffDate) {
-              skippedOldRuns++;
-              totalSkippedOldRuns++;
-              continue;
-            }
-
-            // If we have cached run IDs, check if this run is already cached
-            if (cachedIds.size > 0 && cachedIds.has(runIdStr)) {
-              consecutiveCachedRuns++;
-              skippedCachedRuns++;
-              totalSkippedCachedRuns++;
-              // If we've seen many consecutive cached runs, we've likely reached the boundary
-              if (consecutiveCachedRuns >= MAX_CONSECUTIVE_CACHED) {
-                core.info(`[FETCH] ${workflowId} - Early exit: found ${consecutiveCachedRuns} consecutive cached runs`);
-                shouldStopWorkflow = true;
-                break;
-              }
-              continue;
-            }
-
-            // Reset consecutive cached counter when we find a new run
-            if (consecutiveCachedRuns > 0) {
-              core.info(`[FETCH] ${workflowId} - Found new run after ${consecutiveCachedRuns} cached runs, resetting counter`);
-            }
-            consecutiveCachedRuns = 0;
-
-            // This is a new run, add it
-            addedNewRuns++;
-            totalAddedNewRuns++;
-            allRuns.push(run);
-            if (addedNewRuns <= 5) {
-              core.info(`[FETCH] ${workflowId} - Added new run ${runIdStr} (${run.name}, ${runDate.toISOString()})`);
-            }
-          }
-
-          if (shouldStopWorkflow) {
-            break;
-          }
-
-          // If we got fewer runs than requested, we've reached the end
-          if (runs.workflow_runs.length < RUNS_PER_PAGE) {
-            break;
-          }
-
-          // Small delay between pages to avoid rate limiting
-          await delay(100);
-        } catch (e) {
-          core.warning(`[FETCH] Failed to fetch runs for workflow ${workflowId} page ${page}: ${e.message}`);
-          break;
-        }
-      }
-
-      core.info(`[FETCH] ${workflowId} - Summary: added ${addedNewRuns} new runs, skipped ${skippedCachedRuns} cached, ${skippedOldRuns} old`);
-
-      // Small delay between workflows to avoid rate limiting
-      await delay(200);
-    }
-
-    core.info(`[FETCH] Completed all workflows. Total: added ${totalAddedNewRuns} new runs, skipped ${totalSkippedCachedRuns} cached, ${totalSkippedOldRuns} old`);
-    return allRuns;
-  }
-
-  // Backward compatibility: if no workflow IDs provided, use old behavior
-  core.info(`[FETCH] No workflow IDs provided, using backward compatibility mode (fetching all workflows)`);
-  let consecutiveCachedRuns = 0;
-  let skippedOldRuns = 0;
-  let skippedCachedRuns = 0;
-  let addedNewRuns = 0;
-
-  for (let page = 1; page <= MAX_PAGES; page++) {
-    core.info(`[FETCH] Fetching page ${page}...`);
-    const params = {
-      owner: context.repo.owner,
-      repo: context.repo.repo,
-      per_page: RUNS_PER_PAGE,
-      page,
-    };
-    if (eventType) {
-      params.event = eventType;
-    }
-    const { data: runs } = await github.rest.actions.listWorkflowRunsForRepo(params);
-    if (!runs.workflow_runs.length) {
-      core.info(`[FETCH] No runs on page ${page}, stopping`);
-      break;
-    }
-
-    core.info(`[FETCH] Page ${page}: processing ${runs.workflow_runs.length} runs`);
-
-    for (const run of runs.workflow_runs) {
-      const runDate = new Date(run.created_at);
-      const runIdStr = String(run.id);
-
-      // Early exit if runs are older than 2 weeks
-      if (runDate < twoWeeksAgo) {
-        core.info(`[FETCH] Early exit: runs older than 2 weeks (${runDate.toISOString()} < ${twoWeeksAgo.toISOString()})`);
-        core.info(`[FETCH] Summary: added ${addedNewRuns} new runs, skipped ${skippedCachedRuns} cached, ${skippedOldRuns} old`);
-        return allRuns;
-      }
-
-      // Skip runs older than cutoff date
-      if (runDate < cutoffDate) {
-        skippedOldRuns++;
-        if (skippedOldRuns <= MAX_CONSECUTIVE_CACHED) {
-          core.info(`[FETCH] Skipping run ${runIdStr} (older than cutoff: ${runDate.toISOString()} < ${cutoffDate.toISOString()})`);
-        }
-        continue;
-      }
-
-      // If we have cached run IDs, check if this run is already cached
-      if (cachedIds.size > 0 && cachedIds.has(runIdStr)) {
-        consecutiveCachedRuns++;
-        skippedCachedRuns++;
-        if (consecutiveCachedRuns <= MAX_CONSECUTIVE_CACHED) {
-          core.info(`[FETCH] Skipping cached run ${runIdStr} (consecutive cached: ${consecutiveCachedRuns})`);
-        }
-        // If we've seen many consecutive cached runs, we've likely reached the boundary
-        if (consecutiveCachedRuns >= MAX_CONSECUTIVE_CACHED) {
-          core.info(`[FETCH] Early exit: found ${consecutiveCachedRuns} consecutive cached runs, stopping fetch`);
-          core.info(`[FETCH] Summary: added ${addedNewRuns} new runs, skipped ${skippedCachedRuns} cached, ${skippedOldRuns} old`);
-          return allRuns;
-        }
-        continue;
-      }
-
-      // Reset consecutive cached counter when we find a new run
-      if (consecutiveCachedRuns > 0) {
-        core.info(`[FETCH] Found new run after ${consecutiveCachedRuns} cached runs, resetting counter`);
-      }
-      consecutiveCachedRuns = 0;
-
-      // This is a new run, add it
-      addedNewRuns++;
-      allRuns.push(run);
-      if (addedNewRuns <= 10) {
-        core.info(`[FETCH] Added new run ${runIdStr} (${run.name}, ${runDate.toISOString()})`);
-      }
-    }
-
-    // If the api call returned no runs, assume we've reached the end and break
-    if (!runs.workflow_runs.length) break;
-  }
-
-  core.info(`[FETCH] Completed all pages. Summary: added ${addedNewRuns} new runs, skipped ${skippedCachedRuns} cached, ${skippedOldRuns} old`);
-  return allRuns;
-}
-
-/**
- * Group runs by workflow name.
- * @param {Array} runs - Array of workflow run objects
- * @returns {Map} Map of workflow name to array of runs
- */
-function groupRunsByName(runs) {
-  const grouped = new Map();
-  for (const run of runs) {
-    if (!grouped.has(run.name)) {
-      grouped.set(run.name, []);
-    }
-    grouped.get(run.name).push(run);
-  }
-  return grouped;
-}
-
-/**
- * Find the most recent successful run of aggregate-workflow-data workflow on main branch.
- * @param {object} octokit - Octokit client
- * @param {object} context - GitHub Actions context
- * @returns {Promise<number|null>} Run ID or null if not found
- */
-async function findPreviousAggregateRun(octokit, context, branch = 'main') {
-  try {
-    core.info('[CACHE] Searching for previous successful aggregate-workflow-data run...');
-    const workflowId = '.github/workflows/aggregate-workflow-data.yaml';
-    const resp = await octokit.rest.actions.listWorkflowRuns({
-      owner: context.repo.owner,
-      repo: context.repo.repo,
-      workflow_id: workflowId,
-      branch: branch,
-      status: 'completed',
-      per_page: 10
-    });
-    const allRuns = resp.data.workflow_runs || [];
-    core.info(`[CACHE] Found ${allRuns.length} completed runs for aggregate-workflow-data`);
-    const runs = allRuns.filter(r => r.conclusion === 'success');
-    core.info(`[CACHE] Found ${runs.length} successful runs`);
-    if (runs.length > 0) {
-      core.info(`[CACHE] Selected previous run ID: ${runs[0].id} (created: ${runs[0].created_at})`);
-      return runs[0].id;
-    }
-    core.info('[CACHE] No previous successful run found');
-    return null;
-  } catch (e) {
-    core.warning(`[CACHE] Failed to find previous aggregate run: ${e.message}`);
-    return null;
-  }
-}
-
-/**
- * Prune data older than cutoff date from grouped runs.
- * @param {Map} grouped - Map of workflow name to array of runs
- * @param {Date} cutoffDate - Cutoff date
- * @returns {Map} Pruned grouped runs
- */
-function pruneOldRuns(grouped, cutoffDate) {
-  const pruned = new Map();
-  let totalRemoved = 0;
-  for (const [name, runs] of grouped.entries()) {
-    const beforeCount = runs.length;
-    const filtered = runs.filter(run => {
-      const runDate = new Date(run.created_at);
-      return runDate >= cutoffDate;
-    });
-    const removed = beforeCount - filtered.length;
-    if (removed > 0) {
-      totalRemoved += removed;
-      const newestRemoved = runs.find(r => new Date(r.created_at) < cutoffDate);
-      if (newestRemoved) {
-        core.info(`[PRUNE] Workflow '${name}': removed ${removed} runs (newest removed: ${newestRemoved.created_at})`);
-      }
-    }
-    if (filtered.length > 0) {
-      pruned.set(name, filtered);
-    } else {
-      core.info(`[PRUNE] Workflow '${name}': removed all ${beforeCount} runs (workflow has no runs >= ${cutoffDate.toISOString()})`);
-    }
-  }
-  if (totalRemoved > 0) {
-    core.info(`[PRUNE] Total runs removed across all workflows: ${totalRemoved}`);
-  } else {
-    core.info(`[PRUNE] No runs removed (all runs are >= ${cutoffDate.toISOString()})`);
-  }
-  return pruned;
-}
-
-/**
- * Prune annotations index to remove entries for runs older than cutoff date.
- * @param {object} annotationsIndex - Index mapping runId -> directory
- * @param {Map} grouped - Map of workflow name to array of runs (for date lookup)
- * @param {Date} cutoffDate - Cutoff date
- * @returns {object} Pruned annotations index
- */
-function pruneAnnotationsIndex(annotationsIndex, grouped, cutoffDate) {
-  const pruned = {};
-  const runDates = new Map();
-  // Build map of run IDs to dates
-  for (const runs of grouped.values()) {
-    for (const run of runs) {
-      runDates.set(String(run.id), new Date(run.created_at));
-    }
-  }
-  for (const [runId, dir] of Object.entries(annotationsIndex || {})) {
-    const runDate = runDates.get(runId);
-    if (runDate && runDate >= cutoffDate) {
-      pruned[runId] = dir;
-    }
-  }
-  return pruned;
-}
-
-/**
- * Prune logs index to remove entries for runs older than cutoff date.
- * @param {object} logsIndex - Index mapping runId -> directory
- * @param {Map} grouped - Map of workflow name to array of runs (for date lookup)
- * @param {Date} cutoffDate - Cutoff date
- * @returns {object} Pruned logs index
- */
-function pruneLogsIndex(logsIndex, grouped, cutoffDate) {
-  const pruned = {};
-  const runDates = new Map();
-  // Build map of run IDs to dates
-  for (const runs of grouped.values()) {
-    for (const run of runs) {
-      runDates.set(String(run.id), new Date(run.created_at));
-    }
-  }
-  for (const [runId, dir] of Object.entries(logsIndex || {})) {
-    const runDate = runDates.get(runId);
-    if (runDate && runDate >= cutoffDate) {
-      pruned[runId] = dir;
-    }
-  }
-  return pruned;
-}
-
-/**
- * Prune commits older than cutoff date.
- * @param {Array} commits - Array of commit objects
- * @param {Date} cutoffDate - Cutoff date
- * @returns {Array} Pruned commits
- */
-function pruneCommits(commits, cutoffDate) {
-  return (commits || []).filter(c => {
-    const commitDate = c.date ? new Date(c.date) : null;
-    return commitDate && commitDate >= cutoffDate;
-  });
-}
-
-/**
- * Check if a workflow name matches any configuration in workflow_configs.
- * @param {string} workflowName - Name of the workflow to check
- * @param {Array} workflowConfigs - Array of config objects with wkflw_name or wkflw_prefix
- * @returns {boolean} True if workflow matches any config
- */
-function workflowMatchesConfig(workflowName, workflowConfigs) {
-  if (!Array.isArray(workflowConfigs) || workflowConfigs.length === 0) {
-    return true; // If no configs provided, match all workflows (backward compatibility)
-  }
-  for (const config of workflowConfigs) {
-    if (config.wkflw_name && workflowName === config.wkflw_name) {
-      return true;
-    }
-    if (config.wkflw_prefix && workflowName.startsWith(config.wkflw_prefix)) {
-      return true;
-    }
-  }
-  return false;
-}
-
-/**
- * Download and extract an artifact to a temporary directory.
- * @param {object} octokit - Octokit client
- * @param {string} owner - Repository owner
- * @param {string} repo - Repository name
- * @param {object} artifact - Artifact object
- * @param {string} tmpDir - Temporary directory path
- * @returns {Promise<string>} Path to extracted directory
- */
-async function downloadAndExtractArtifact(octokit, owner, repo, artifact, tmpDir) {
-  const resp = await octokit.rest.actions.downloadArtifact({ owner, repo, artifact_id: artifact.id, archive_format: 'zip' });
-  const zipPath = path.join(tmpDir, `${artifact.name}.zip`);
-  fs.writeFileSync(zipPath, Buffer.from(resp.data));
-  const extractDir = path.join(tmpDir, `${artifact.name}-extract`);
-  if (!fs.existsSync(extractDir)) fs.mkdirSync(extractDir, { recursive: true });
-  execFileSync('unzip', ['-o', zipPath, '-d', extractDir], { stdio: 'ignore' });
-  return extractDir;
-}
-
-/**
- * Recursively find a file by name in a directory.
- * @param {string} rootDir - Root directory to search
- * @param {string|Set<string>} targetFileName - File name(s) to find
- * @returns {string|null} Path to found file or null
- */
-function findFileInDirectory(rootDir, targetFileName) {
-  const targetNames = typeof targetFileName === 'string' ? new Set([targetFileName]) : targetFileName;
-  const stack = [rootDir];
-  while (stack.length) {
-    const dir = stack.pop();
-    const entries = fs.readdirSync(dir, { withFileTypes: true });
-    for (const ent of entries) {
-      const p = path.join(dir, ent.name);
-      if (ent.isDirectory()) {
-        stack.push(p);
-      } else if (ent.isFile() && targetNames.has(ent.name)) {
-        return p;
-      }
-    }
-  }
-  return null;
-}
-
-/**
- * Remove directories in a root directory that are not in the kept set.
- * @param {string} rootDir - Root directory to clean
- * @param {Set<string>} keptIds - Set of directory names to keep
- * @param {string} indexFileName - Name of index file to exclude from removal
- * @returns {number} Number of directories removed
- */
-function removeUnkeptDirectories(rootDir, keptIds, indexFileName) {
-  if (!fs.existsSync(rootDir)) {
-    return 0;
-  }
-  let removedDirs = 0;
-  const entries = fs.readdirSync(rootDir, { withFileTypes: true });
-  for (const ent of entries) {
-    if (ent.isDirectory() && ent.name !== indexFileName) {
-      if (!keptIds.has(ent.name)) {
-        fs.rmSync(path.join(rootDir, ent.name), { recursive: true, force: true });
-        removedDirs++;
-      }
-    }
-  }
-  return removedDirs;
-}
 
 /**
  * Main entrypoint for the action.
