+++ .github/actions/fetch-workflow-data/lib/data-building.js
@@ -1,0 +1,7 @@
+// Data Building Module
+// Handles commits index building and last success timestamp tracking
+// Note: Functions will be extracted from run() in a later phase
+
+module.exports = {
+  // Functions will be added here when extracted from run()
+};
